app/requirements.txt
uvicorn
fastapi
pydantic

singleton_package
json-advanced
python-dotenv
debugpy
ipython

aiohttp
aiofiles
aiocache
usso[fastapi]

beanie
fastapi-mongo-base
ufaas-fastapi-business
httpx